% --- gada-internal.Rd ---

\name{gada-internal}
\alias{gada-internal}
\alias{collapseInfo}
\alias{WextIextToSegments}
\alias{SBL.fit}
\alias{BackwardElimination.fit}
\alias{findNormalLimits}

\title{Internal gada functions}
\description{Internal gada functions}
\usage{
collapseInfo(x)
WextIextToSegments(x)
SBL.fit(i,x,control,gen.info,chr,sigma2,aAlpha,maxit,tol,debug)
BackwardElimination.fit(i,x,chr,T,MinSegLen)
findNormalLimits(x)
}

\details{These are not to be called by the user}
\keyword{internal}


